LDE-714: moved read tests
